backend/requirements.txt
# B2 Second Brain - Backend Dependencies
# Python 3.10+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Caching (in-memory, replaces Redis)
cachetools==5.3.2

# AI / ML
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.14.0
langchain==0.1.4
langchain-community==0.0.14
langchain-core>=0.1.16,<0.2

# Embeddings
sentence-transformers==2.3.1
torch>=2.0.0
transformers==4.37.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0
# paddleocr==2.7.0  # Optional - install if needed

# Speech-to-Text
openai-whisper==20231117
# faster-whisper==0.10.0  # Optional faster alternative

# Document Processing
PyMuPDF==1.23.22
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
markdown==3.5.2
beautifulsoup4==4.12.3
lxml==4.9.4
chardet==5.2.0

# Web Scraping
trafilatura==1.6.4
newspaper3k==0.2.8
youtube-transcript-api==0.6.2

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
aiofiles==23.2.1
pytz==2024.1
tqdm==4.66.1

# Code Parsing
pygments==2.17.2
tree-sitter==0.20.4

# Data Validation
email-validator==2.1.0.post1
validators==0.22.0

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development Tools (dev)
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7
